+++ src/main/kotlin/com/derek/backuptool/service/PrimaryBackupService.kt
@@ -1,0 +1,13 @@
+package com.derek.backuptool.service
+
+import com.derek.backuptool.dto.BackupToolConfig
+import com.derek.backuptool.service.BackupToolService
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class PrimaryBackupService @Inject constructor() : BackupToolService {
+    override fun performService(backupToolConfig: BackupToolConfig) {
+        TODO("Not yet implemented")
+    }
+}
