+++ src/main/kotlin/com/derek/backuptool/service/S3BackupService.kt
@@ -1,11 +1,10 @@
-package com.derek.backuptool.service.backuptoolservice
+package com.derek.backuptool.service
 
 import com.derek.backuptool.dto.BackupToolConfig
+import com.derek.backuptool.service.BackupToolService
 import javax.inject.Inject
-import javax.inject.Singleton
 
-@Singleton
-class PrimaryBackupService @Inject constructor() : BackupToolService {
+class S3BackupService @Inject constructor() : BackupToolService {
     override fun performService(backupToolConfig: BackupToolConfig) {
         TODO("Not yet implemented")
     }
