+++ src/main/kotlin/com/derek/backuptool/service/StatisticsService.kt
@@ -1,9 +1,10 @@
-package com.derek.backuptool.service.backuptoolservice
+package com.derek.backuptool.service
 
 import com.derek.backuptool.dto.BackupToolConfig
+import com.derek.backuptool.service.BackupToolService
 import javax.inject.Inject
 
-class S3BackupService @Inject constructor() : BackupToolService {
+class StatisticsService @Inject constructor() : BackupToolService {
     override fun performService(backupToolConfig: BackupToolConfig) {
         TODO("Not yet implemented")
     }
