--- src/main/kotlin/com/derek/backuptool/service/backuptoolservice/StatisticsService.kt
@@ -1,10 +1,0 @@
-package com.derek.backuptool.service.backuptoolservice
-
-import com.derek.backuptool.dto.BackupToolConfig
-import javax.inject.Inject
-
-class StatisticsService @Inject constructor() : BackupToolService {
-    override fun performService(backupToolConfig: BackupToolConfig) {
-        TODO("Not yet implemented")
-    }
-}
